requirements.txt
beautifulsoup4==4.6.0
boto3==1.9.199
botocore==1.12.199
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.3
django-amp-tools==0.2.2
django-colorful==1.3
django-el-pagination==3.2.4
django-heroku==0.3.1
django-modelcluster==4.4
django-redis==4.10.0
django-social-share==1.3.2
django-storages==1.7.1
django-taggit==0.24.0
django-toolbelt==0.0.1
django-treebeard==4.3
djangorestframework==3.10.1
docutils==0.14
draftjs-exporter==2.1.6
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
Pillow==5.4.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
pytz==2019.1
redis==3.3.4
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.5.1
wagtail-cache==0.5.2
webencodings==0.5.1
whitenoise==3.0
Willow==1.1
